// types/destination.ts
export type Destination = {
  name: string
  slug: string
  city: string
  category: string
  content: string
  mainImage: string
  michelinStars: number
  crown: boolean
  brand?: string
  cardTags?: string
  lat?: number
  long?: number
  myRating?: number
  reviewed?: boolean
  subline?: string
}